Trust users from given teams or trusted list
This allows finer control of the RequiredReviewApprovals property. It is
ignored if the user is in one of the trusted teams, or in the trusted
user list.

This way we can run CI on trusted users without needing an approval, but
untrusted users will require a PR approval before their changes are run
in CI.

In order to reduce the number of calls to github we are caching trusted
users for the duration of a single check.

Note: github should be configured to remove approvals if new commits are
received on a PR so that a user doesn't circumvent security by pushing a
malicious commit after an approved legitimate commit on the same PR.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>